requirements.txt
faiss-cpu==1.7.4
faiss-gpu==1.7.2
sentence_transformers==3.0.1
onnxruntime==1.14.1
onnxruntime-gpu==1.14.1
langchain==0.1.13
einops
pandas
transformers==4.38.1
dataclasses_json
fire
scikit-learn
ipdb
html2text
markdown
beautifulsoup4
dill
openpyxl
fuzzywuzzy
fastapi
torch==2.0.0
flask[async]
gunicorn
loguru
langchain_community
rank_bm25
accelerate==0.28.0
spacy==3.7.4
nltk==3.8.1
deepspeed==0.14.0